fix warnings in CFeatureExtraction

+++ libs/vision/src/CFeatureExtraction_LSD_BLD.cpp
@@ -74,9 +74,9 @@ void CFeatureExtraction::extractFeaturesLSD(
 	// Now:
 	//  1) Sort them by "response":
 	// sort the LSD features by line length
-	for (int i = 0; i < N; i++)
+	for (size_t i = 0; i < N; i++)
 	{
-		for (int j = i + 1; j < N; j++)
+		for (size_t j = i + 1; j < N; j++)
 		{
 			if (cv_line.at(j).lineLength > cv_line.at(i).lineLength)
 			{
